Create 1400. Construct K Palindrome Strings
Initialization:

Check if the length of the string s is less than k. If it is, return false since it's impossible to construct k palindrome strings.

Create a HashMap to count the frequency of each character in the string s.

Counting Characters:

Iterate through the string s and update the HashMap with the frequency of each character.

Counting Odd Frequencies:

Initialize a variable oddCount to count the number of characters with odd frequencies.

Iterate through the values in the HashMap and increment oddCount for each character with an odd frequency.

Checking Palindrome Construction:

Return true if the number of characters with odd frequencies (oddCount) is less than or equal to k. This is because each palindrome can have at most one character with an odd frequency, and we can distribute the odd frequency characters among the k palindromes.